Switch to override, update doi addition, and fix test file name
